% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addLogn}
\alias{addLogn}
\title{Apply lognormal observation error}
\usage{
addLogn(x, cv)
}
\arguments{
\item{x}{Numeric vector.}

\item{cv}{Coefficient of variation.}
}
\value{
Perturbed numeric vector.
}
\description{
Apply lognormal observation error
}
