% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{addLognormalNoise}
\alias{addLognormalNoise}
\title{Add low lognormal noise}
\usage{
addLognormalNoise(x, sdlog = 0.02)
}
\arguments{
\item{x}{Numeric vector.}

\item{sdlog}{Log-scale standard deviation.}
}
\value{
Numeric vector with multiplicative noise.
}
\description{
Add low lognormal noise
}
