% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{asTruth}
\alias{asTruth}
\alias{asTruth,list-method}
\alias{asTruth,ANY-method}
\title{Convert SS object to standardized truth table}
\usage{
asTruth(x, ...)

\S4method{asTruth}{list}(x, ...)

\S4method{asTruth}{ANY}(x, ...)
}
\arguments{
\item{x}{SS truth object.}

\item{...}{Reserved for future method-specific options.}
}
\value{
Data frame with \code{year}, \code{catch}, \code{stock}, and \code{ssb}.
}
\description{
Convert SS object to standardized truth table
}
