% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{catchDat}
\alias{catchDat}
\title{Reported catch total from data file object for one year}
\usage{
catchDat(dat, year, nfish)
}
\arguments{
\item{dat}{List from \code{SS_readdat()}.}

\item{year}{Calendar year.}

\item{nfish}{Number of fishing fleets (non-survey).}
}
\description{
Reported catch total from data file object for one year
}
