% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{catchMsy}
\alias{catchMsy}
\title{Dead catch relative to MSY benchmark}
\usage{
catchMsy(rep, deadCatchTotal)
}
\arguments{
\item{rep}{A list returned by \code{SS_output()}.}

\item{deadCatchTotal}{Total dead catch (same units as \code{Dead_Catch_MSY}).}
}
\description{
Dead catch relative to MSY benchmark
}
