% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{compareCurrent}
\alias{compareCurrent}
\title{Compare SS truth and JABBA current state in terminal year}
\usage{
compareCurrent(truthDf, jabbaFit = NULL, jabbaInput = NULL)
}
\arguments{
\item{truthDf}{Data frame from \code{asTruth()}.}

\item{jabbaFit}{JABBA fit object or fit specification.}

\item{jabbaInput}{JABBA input list.}
}
\value{
One-row data frame with terminal status metrics.
}
\description{
Compare SS truth and JABBA current state in terminal year
}
