% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{compareHistorical}
\alias{compareHistorical}
\title{Compare SS truth and JABBA historical trajectories}
\usage{
compareHistorical(truthDf, jabbaFit = NULL, jabbaInput = NULL)
}
\arguments{
\item{truthDf}{Data frame from \code{asTruth()}.}

\item{jabbaFit}{JABBA fit object or fit specification.}

\item{jabbaInput}{JABBA input list.}
}
\value{
Data frame comparing observed trajectories and relative scales.
}
\description{
Compare SS truth and JABBA historical trajectories
}
