% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-run.R
\name{copySsInputs}
\alias{copySsInputs}
\title{Copy SS3 input files into a destination directory}
\usage{
copySsInputs(from_dir, to_dir)
}
\arguments{
\item{from_dir}{Source scenario directory.}

\item{to_dir}{Destination directory (created if needed).}
}
\description{
Copy SS3 input files into a destination directory
}
