% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-ss-utils.R
\name{curveSsNeedsEnrichment}
\alias{curveSsNeedsEnrichment}
\title{Check whether a curveSS object needs FLRebuild enrichment for PE plots}
\usage{
curveSsNeedsEnrichment(pe)
}
\arguments{
\item{pe}{List returned by \code{\link{curveSS}}.}
}
\description{
Check whether a curveSS object needs FLRebuild enrichment for PE plots
}
