% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-ss-utils.R
\name{curveSsProductionCol}
\alias{curveSsProductionCol}
\title{Resolve production column name in curveSS tseries}
\usage{
curveSsProductionCol(ts)
}
\arguments{
\item{ts}{\code{tseries} data frame.}
}
\description{
Resolve production column name in curveSS tseries
}
