% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{deadB}
\alias{deadB}
\title{Sum dead(B) catch columns for one timeseries row}
\usage{
deadB(tsdf, rowIdx, rep)
}
\arguments{
\item{tsdf}{Timeseries data.frame.}

\item{rowIdx}{Row index.}

\item{rep}{A list returned by \code{SS_output()} (for fleet count).}
}
\description{
Sum dead(B) catch columns for one timeseries row
}
