% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-diags.R
\name{diagSS}
\alias{diagSS}
\title{Extract standardized diagnostics from Stock Synthesis output}
\usage{
diagSS(x, ...)
}
\arguments{
\item{x}{SS3 model directory.}

\item{...}{Reserved.}
}
\value{
Standardized residual diagnostics data frame.
}
\description{
Extract standardized diagnostics from Stock Synthesis output
}
