% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{dqVal}
\alias{dqVal}
\title{Single value from SS3 derived quantities}
\usage{
dqVal(rep, label)
}
\arguments{
\item{rep}{A list returned by \code{SS_output()}.}

\item{label}{Row label in \code{derived_quants}.}
}
\description{
Single value from SS3 derived quantities
}
