% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-ensemble.R
\name{ensembleWeightsEqual}
\alias{ensembleWeightsEqual}
\title{Equal weights for \code{n} ensemble members}
\usage{
ensembleWeightsEqual(n)
}
\arguments{
\item{n}{Number of models.}
}
\value{
Numeric vector of length \code{n} summing to 1.
}
\description{
Equal weights for \code{n} ensemble members
}
