% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{fFmsy}
\alias{fFmsy}
\title{F / F_MSY style ratio from derived quantities}
\usage{
fFmsy(rep, yr)
}
\arguments{
\item{rep}{A list returned by \code{SS_output()}.}

\item{yr}{Calendar year.}
}
\description{
Uses \eqn{F_Y \times}{F_Y x} annF_SPR / annF_MSY.
}
