% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbSs3.R
\name{findCol}
\alias{findCol}
\title{Helper to find the first matching column name}
\usage{
findCol(x, candidates)
}
\arguments{
\item{x}{data.frame}

\item{candidates}{character vector of possible column names}
}
\value{
column name or NA_character_
}
\description{
Helper to find the first matching column name
}
