% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbSs3.R
\name{findObj}
\alias{findObj}
\title{Helper to find the first matching object name in SS_output}
\usage{
findObj(ss3, candidates)
}
\arguments{
\item{ss3}{SS_output list}

\item{candidates}{character vector of possible object names}
}
\value{
object name or NA_character_
}
\description{
Helper to find the first matching object name in SS_output
}
