% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{fleetF}
\alias{fleetF}
\title{Sum fleet F columns for one timeseries row}
\usage{
fleetF(tsdf, rowIdx)
}
\arguments{
\item{tsdf}{Timeseries data.frame.}

\item{rowIdx}{Row index.}
}
\description{
Sum fleet F columns for one timeseries row
}
