% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbSs3.R
\name{getSpawnSeas}
\alias{getSpawnSeas}
\title{Get spawning season from SS3 output}
\usage{
getSpawnSeas(ss3)
}
\arguments{
\item{ss3}{List from \code{r4ss::SS_output()}.}
}
\value{
integer spawning season
}
\description{
Get spawning season from SS3 output
}
