% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mergeLists}
\alias{mergeLists}
\title{Merge two named lists}
\usage{
mergeLists(x, y)
}
\arguments{
\item{x}{Base list.}

\item{y}{List with overriding values.}
}
\value{
Merged list.
}
\description{
Merge two named lists
}
