% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-profile-beta-grid.R
\name{patchCtlSrSurvBeta}
\alias{patchCtlSrSurvBeta}
\title{Set SR_surv_Beta in a Stock Synthesis control file}
\usage{
patchCtlSrSurvBeta(dir, ctlfile, beta)
}
\arguments{
\item{dir}{Scenario directory.}

\item{ctlfile}{Control file name.}

\item{beta}{SR_surv_Beta value.}
}
\description{
Set SR_surv_Beta in a Stock Synthesis control file
}
