% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{peelOne}
\alias{peelOne}
\title{One-step-ahead peel metrics vs base assessment}
\usage{
peelOne(repPeel, repBase, dat, peelLabel, pathPeel)
}
\arguments{
\item{repPeel}{\code{SS_output()} for the peel.}

\item{repBase}{\code{SS_output()} for the base (\code{retro0}).}

\item{dat}{Data list from \code{SS_readdat()} for the scenario.}

\item{peelLabel}{Character peel label.}

\item{pathPeel}{Path to peel directory (for errors).}
}
\description{
One-step-ahead peel metrics vs base assessment
}
