% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{plotAll}
\alias{plotAll}
\title{Build all SS-vs-JABBA comparison plots using ggplot2}
\usage{
plotAll(comparisons)
}
\arguments{
\item{comparisons}{Named list from \code{compareAll()} or \code{workflow()$comparisons}.}
}
\value{
Named list of ggplot objects.
}
\description{
Build all SS-vs-JABBA comparison plots using ggplot2
}
