% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbPlot.R
\name{plotCatchSex}
\alias{plotCatchSex}
\title{Plot sex-specific SSB-equivalent catch over time}
\usage{
plotCatchSex(spSex)
}
\arguments{
\item{spSex}{List from \code{spSsbBySexFromSs3()}.}
}
\value{
A ggplot object showing female vs male catchSsb by year.
}
\description{
Plot sex-specific SSB-equivalent catch over time
}
