% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{plotCurrent}
\alias{plotCurrent}
\title{Plot current-state comparison using ggplot2}
\usage{
plotCurrent(current_state)
}
\arguments{
\item{current_state}{Current-state table from \code{compareCurrent()}.}
}
\value{
ggplot object.
}
\description{
Plot current-state comparison using ggplot2
}
