% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{plotHistorical}
\alias{plotHistorical}
\title{Plot historical trend comparison using ggplot2}
\usage{
plotHistorical(historical)
}
\arguments{
\item{historical}{Historical table from \code{compareHistorical()}.}
}
\value{
ggplot object.
}
\description{
Plot historical trend comparison using ggplot2
}
