% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{plotProduction}
\alias{plotProduction}
\title{Plot production function comparison using ggplot2}
\usage{
plotProduction(production)
}
\arguments{
\item{production}{Production table from \code{compareProduction()}.}
}
\value{
ggplot object.
}
\description{
Plot production function comparison using ggplot2
}
