% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{plotProjection}
\alias{plotProjection}
\title{Plot projections (F=0 and F=FMSY) using ggplot2}
\usage{
plotProjection(projections)
}
\arguments{
\item{projections}{Projection table from \code{compareProjection()}.}
}
\value{
ggplot object.
}
\description{
Plot projections (F=0 and F=FMSY) using ggplot2
}
