% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-diags.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot residual time series with ggplot2}
\usage{
plotResiduals(x, by = "name")
}
\arguments{
\item{x}{Diagnostics data frame (typically from \code{diags()}).}

\item{by}{Optional grouping column for facets.}
}
\value{
ggplot object.
}
\description{
Plot residual time series with ggplot2
}
