% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-diags.R
\name{plotRunTest}
\alias{plotRunTest}
\title{Plot runs-style diagnostics with ggplot2}
\usage{
plotRunTest(x, by = "name")
}
\arguments{
\item{x}{Diagnostics data frame (typically from \code{diags()}).}

\item{by}{Optional grouping column for facets.}
}
\value{
ggplot object.
}
\description{
Plot runs-style diagnostics with ggplot2
}
