% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbPlot.R
\name{plotSpCompare}
\alias{plotSpCompare}
\title{Plot naive vs SSB-based surplus production}
\usage{
plotSpCompare(spDf)
}
\arguments{
\item{spDf}{Data.frame with year, ssb, spSsb, and spNaive.}
}
\value{
A ggplot object.
}
\description{
Plot naive vs SSB-based surplus production
}
