% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbPlot.R
\name{plotSpSex}
\alias{plotSpSex}
\title{Plot sex-specific SP vs SSB}
\usage{
plotSpSex(spSex, addSmooth = TRUE)
}
\arguments{
\item{spSex}{List from \code{spSsbBySexFromSs3()}.}

\item{addSmooth}{Logical; if TRUE, add smooth curves.}
}
\value{
A ggplot object with panels for each sex.
}
\description{
Plot sex-specific SP vs SSB
}
