% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbPlot.R
\name{plotSpSsb}
\alias{plotSpSsb}
\title{Plot surplus production vs SSB}
\usage{
plotSpSsb(spDf, addSmooth = TRUE, showYears = FALSE)
}
\arguments{
\item{spDf}{Data.frame from \code{spSsbFromSs3()}.}

\item{addSmooth}{Logical; if TRUE, add a smooth curve.}

\item{showYears}{Logical; if TRUE, label points by year.}
}
\value{
A ggplot object.
}
\description{
Plot surplus production vs SSB
}
