% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-diags.R
\name{plotStdResiduals}
\alias{plotStdResiduals}
\title{Plot standardized residual time series with ggplot2}
\usage{
plotStdResiduals(x, by = "name")
}
\arguments{
\item{x}{Diagnostics data frame (typically from \code{diags()}).}

\item{by}{Optional grouping column for facets.}
}
\value{
ggplot object.
}
\description{
Plot standardized residual time series with ggplot2
}
