% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-profile-beta-grid.R
\name{profileBetaGrid}
\alias{profileBetaGrid}
\title{Build an M x SR_surv_Beta profile grid}
\usage{
profileBetaGrid(M, beta)
}
\arguments{
\item{M}{Numeric vector of natural-mortality multipliers.}

\item{beta}{Numeric vector of \code{SR_surv_Beta} values.}
}
\description{
Build an M x SR_surv_Beta profile grid
}
