% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-likelihood.R
\name{profileLikelihoodExclude}
\alias{profileLikelihoodExclude}
\title{Likelihood components excluded from profile component plots}
\usage{
profileLikelihoodExclude()
}
\description{
Likelihood components excluded from profile component plots
}
