% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{readWarns}
\alias{readWarns}
\title{Read non-empty warning lines from file}
\usage{
readWarns(path)
}
\arguments{
\item{path}{Path to text file.}
}
\value{
Character vector of non-empty lines.
}
\description{
Read non-empty warning lines from file
}
