% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{rescaleComp}
\alias{rescaleComp}
\title{Rescale composition vector to sum to one}
\usage{
rescaleComp(x)
}
\arguments{
\item{x}{Numeric composition vector.}
}
\value{
Rescaled vector (unchanged when sum is not positive).
}
\description{
Rescale composition vector to sum to one
}
