% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{retroTag}
\alias{retroTag}
\title{Subdirectory name for a retrospective peel}
\usage{
retroTag(yr)
}
\arguments{
\item{yr}{Peel index (non-positive integer; \code{0} is \code{retro0}).}
}
\description{
Subdirectory name for a retrospective peel
}
