% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{safeNrow}
\alias{safeNrow}
\title{Safe row counter}
\usage{
safeNrow(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Number of rows for data frames, otherwise \code{0}.
}
\description{
Safe row counter
}
