% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-profile-beta-grid.R
\name{setupSs3ProfileScenario}
\alias{setupSs3ProfileScenario}
\title{Copy and patch one M x beta profile scenario}
\usage{
setupSs3ProfileScenario(src_dir, dest_dir, M, beta, ctlfile = "control.ss")
}
\description{
Copy and patch one M x beta profile scenario
}
