% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{sliceJabba}
\alias{sliceJabba}
\alias{sliceJabba,character-method}
\title{Run minimal JABBA vertical slice for sma-natl Scenario-1}
\usage{
sliceJabba(x, ...)

\S4method{sliceJabba}{character}(x, runFit = FALSE, cv = 0.2, ...)
}
\arguments{
\item{x}{SS3 directory path.}

\item{...}{Additional args passed to fit function.}

\item{runFit}{Logical; if \code{TRUE}, attempts to run JABBA fit.}

\item{cv}{Index CV used in JABBA translation.}
}
\value{
List with truth object, truth table, JABBA input, and fit/spec.
}
\description{
Run minimal JABBA vertical slice for sma-natl Scenario-1
}
