% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{smoothSeries}
\alias{smoothSeries}
\title{Smooth a numeric time series}
\usage{
smoothSeries(
  year,
  value,
  method = c("ma", "loess", "none")[1],
  loess_span = 0.3,
  ma_k = 3
)
}
\arguments{
\item{year}{Numeric year/index vector.}

\item{value}{Numeric values.}

\item{method}{Smoothing method (\code{"ma"}, \code{"loess"}, \code{"none"}).}

\item{loess_span}{Loess span.}

\item{ma_k}{Moving-average window size.}
}
\value{
Smoothed numeric vector.
}
\description{
Smooth a numeric time series
}
