% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{spawnIx}
\alias{spawnIx}
\title{Row index in timeseries for spawn season and era}
\usage{
spawnIx(rep, yr, era)
}
\arguments{
\item{rep}{A list returned by \code{SS_output()}.}

\item{yr}{Calendar year.}

\item{era}{\code{"TIME"} or \code{"FORE"}.}
}
\description{
Row index in timeseries for spawn season and era
}
