% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbSs3.R
\name{spSsbBySexFromSs3}
\alias{spSsbBySexFromSs3}
\title{Sex-specific SSB-based surplus production from SS3}
\usage{
spSsbBySexFromSs3(ss3, spawnSeas = getSpawnSeas(ss3))
}
\arguments{
\item{ss3}{List from \code{r4ss::SS_output()}.}

\item{spawnSeas}{Spawning season index.}
}
\value{
List with female and male data.frames.
}
\description{
Sex-specific SSB-based surplus production from SS3
}
