% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbSs3.R
\name{spSsbFromSs3}
\alias{spSsbFromSs3}
\title{SSB-based surplus production time series from SS3}
\usage{
spSsbFromSs3(ss3, sex = 1, spawnSeas = getSpawnSeas(ss3))
}
\arguments{
\item{ss3}{List from \code{r4ss::SS_output()}.}

\item{sex}{Integer (1 = female, 2 = male).}

\item{spawnSeas}{Spawning season index.}
}
\value{
data.frame with year, ssb, catchSsb, spSsb.
}
\description{
SSB-based surplus production time series from SS3
}
