% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{ssAsJabba}
\alias{ssAsJabba}
\alias{ssAsJabba,data.frame-method}
\title{Translate truth table to JABBA input list}
\usage{
ssAsJabba(x, ...)

\S4method{ssAsJabba}{data.frame}(x, indexCol = "ssb", cv = 0.2, ...)
}
\arguments{
\item{x}{Truth data frame from \code{ssAsTruth()}.}

\item{...}{Reserved.}

\item{indexCol}{Optional column name to use as index (defaults to \code{ssb}).}

\item{cv}{Optional scalar CV used to construct index SE.}
}
\value{
Named list with \code{catch}, \code{cpue}, \code{se}, and metadata.
}
\description{
Translate truth table to JABBA input list
}
