% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{ssbOver0}
\alias{ssbOver0}
\title{SSB relative to virgin biomass}
\usage{
ssbOver0(rep, yr)
}
\arguments{
\item{rep}{A list returned by \code{SS_output()}.}

\item{yr}{Calendar year.}
}
\description{
SSB relative to virgin biomass
}
