% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssCatch}
\alias{ssCatch}
\title{Modify catch series with smoothing and noise}
\usage{
ssCatch(
  dat,
  catch_cv = 0.02,
  smooth_method = c("ma", "loess", "none")[1],
  loess_span = 0.3,
  ma_k = 3,
  fleets_smooth = NULL,
  fleets_keep_exact = NULL,
  years_keep_exact = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{SS data list.}

\item{catch_cv}{CV for catch perturbation.}

\item{smooth_method}{One of \code{"ma"}, \code{"loess"}, \code{"none"}.}

\item{loess_span}{Loess span.}

\item{ma_k}{Moving-average window size.}

\item{fleets_smooth}{Fleets eligible for smoothing.}

\item{fleets_keep_exact}{Fleets excluded from changes.}

\item{years_keep_exact}{Years excluded from changes.}

\item{verbose}{Logical flag.}
}
\value{
List with updated \code{dat} and \code{audit}.
}
\description{
Modify catch series with smoothing and noise
}
