% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbSs3.R
\name{ssCatchAge}
\alias{ssCatchAge}
\title{Catch biomass at age from SS3 output}
\usage{
ssCatchAge(ss3, sex = 1, spawnSeas = getSpawnSeas(ss3))
}
\arguments{
\item{ss3}{List from \code{r4ss::SS_output()}.}

\item{sex}{Integer (1 = female, 2 = male).}

\item{spawnSeas}{Spawning season index.}
}
\value{
List with years, ages, catch matrix (year x age).
}
\description{
Catch biomass at age from SS3 output
}
