% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{ssCompareHistorical}
\alias{ssCompareHistorical}
\title{Compare SS truth and JABBA historical trajectories}
\usage{
ssCompareHistorical(truthDf, jabbaFit = NULL, jabbaInput = NULL)
}
\arguments{
\item{truthDf}{Data frame from \code{ssAsTruth()}.}

\item{jabbaFit}{JABBA fit object or fit specification.}

\item{jabbaInput}{JABBA input list.}
}
\value{
Data frame comparing observed trajectories and relative scales.
}
\description{
Compare SS truth and JABBA historical trajectories
}
