% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssComps}
\alias{ssComps}
\title{Inflate composition sample sizes}
\usage{
ssComps(dat, min_comp_n = 500, fleets_keep_exact = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{SS data list from \code{r4ss::SS_readdat()}.}

\item{min_comp_n}{Minimum effective sample size.}

\item{fleets_keep_exact}{Fleets excluded from inflation.}

\item{verbose}{Logical flag.}
}
\value{
List with updated \code{dat} and audit table.
}
\description{
Inflate composition sample sizes
}
