% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssCopyTree}
\alias{ssCopyTree}
\title{Copy a full model directory tree}
\usage{
ssCopyTree(src_dir, dst_dir, overwrite = FALSE)
}
\arguments{
\item{src_dir}{Source directory.}

\item{dst_dir}{Destination directory.}

\item{overwrite}{Logical; overwrite destination if it exists.}
}
\value{
Normalized destination path.
}
\description{
Copy a full model directory tree
}
