% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-export.R
\name{ssCpueFit}
\alias{ssCpueFit}
\title{CPUE fits and residuals from \code{SS_output}}
\usage{
ssCpueFit(rep)
}
\arguments{
\item{rep}{\code{SS_output} list.}
}
\description{
CPUE fits and residuals from \code{SS_output}
}
