% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-export.R
\name{ssCpueInput}
\alias{ssCpueInput}
\title{CPUE input series from \code{SS_output}}
\usage{
ssCpueInput(rep)
}
\arguments{
\item{rep}{\code{SS_output} list.}
}
\description{
CPUE input series from \code{SS_output}
}
