% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssIndex}
\alias{ssIndex}
\title{Apply low-error updates to CPUE/index data}
\usage{
ssIndex(
  dat,
  index_cv = 0.05,
  fleets_keep_exact = NULL,
  fleets_drop = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{SS data list from \code{r4ss::SS_readdat()}.}

\item{index_cv}{CV used for index perturbation.}

\item{fleets_keep_exact}{Fleets excluded from perturbation.}

\item{fleets_drop}{Fleets excluded entirely from modification.}

\item{verbose}{Logical flag.}
}
\value{
List with updated \code{dat} and \code{audit}.
}
\description{
Apply low-error updates to CPUE/index data
}
