% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-export.R
\name{ssLenFromRep}
\alias{ssLenFromRep}
\title{Length-composition slots from \code{SS_output}}
\usage{
ssLenFromRep(rep)
}
\arguments{
\item{rep}{\code{SS_output} list.}
}
\description{
Length-composition slots from \code{SS_output}
}
