% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-read.R
\name{ssLoad}
\alias{ssLoad}
\title{Load \code{SS_output} objects for many runs}
\usage{
ssLoad(
  runs,
  cache = FALSE,
  covar = FALSE,
  forecast = FALSE,
  refresh = FALSE,
  parallel = TRUE,
  workers = NULL
)
}
\arguments{
\item{runs}{\code{ssRuns()} table.}

\item{cache}{If \code{TRUE}, require cached output (use after \code{SS_outputs()}).}
}
\description{
Load \code{SS_output} objects for many runs
}
