% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssModelInfo}
\alias{ssModelInfo}
\title{Summarize SS model input structure}
\usage{
ssModelInfo(ssDir, version = "3.30", verbose = TRUE)
}
\arguments{
\item{ssDir}{SS3 model directory.}

\item{version}{SS version string.}

\item{verbose}{Logical flag.}
}
\value{
List with \code{summary} and \code{notes} data frames.
}
\description{
Summarize SS model input structure
}
