% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-PE.R
\name{ssPe}
\alias{ssPe}
\alias{ssPe,character-method}
\alias{ssPe,list-method}
\title{Surplus-production and PE comparison from SS3 model output}
\usage{
ssPe(x, ...)

\S4method{ssPe}{character}(x, ...)

\S4method{ssPe}{list}(x, ...)
}
\arguments{
\item{x}{Either an SS3 directory (\code{character}) or a \code{list} with elements
\code{tseries} and \code{curve}.}

\item{...}{Additional arguments passed to \code{\link[=ssPeCompare]{ssPeCompare()}}.}
}
\value{
The same list returned by \code{\link[=ssPeCompare]{ssPeCompare()}}.
}
\description{
S4 generic wrapper around \code{\link[=ssPeCompare]{ssPeCompare()}}. For \code{character} input,
\code{x} is interpreted as an SS3 directory and \code{curveSS()} is used to build the
required time-series and equilibrium objects.
}
