% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-PE.R
\name{ssPeDiagnose}
\alias{ssPeDiagnose}
\alias{ssPeDiagnose,character-method}
\alias{ssPeDiagnose,list-method}
\title{Fast diagnostics-only SS PE generic}
\usage{
ssPeDiagnose(x, ...)

\S4method{ssPeDiagnose}{character}(x, ...)

\S4method{ssPeDiagnose}{list}(x, ...)
}
\arguments{
\item{x}{Either an SS3 directory (\code{character}) or a \code{list} with \code{tseries}
and \code{curve}.}

\item{...}{Additional arguments passed to \code{\link[=ssPe]{ssPe()}}.}
}
\value{
Named list of diagnostics metrics.
}
\description{
Returns only the \code{diagnostics} component from \code{\link[=ssPe]{ssPe()}}.
}
