% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-ssbSs3.R
\name{ssPhiYear}
\alias{ssPhiYear}
\title{SSB-equivalent factors for one year}
\usage{
ssPhiYear(lh, yearIndex)
}
\arguments{
\item{lh}{Life-history list from \code{ssLifeHistory()}.}

\item{yearIndex}{Integer index into lh$years.}
}
\value{
List with ages, surv, phi, SBPR, Z and F.
}
\description{
SSB-equivalent factors for one year
}
