% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{ssPlotCurrentState}
\alias{ssPlotCurrentState}
\title{Plot current-state comparison using ggplot2}
\usage{
ssPlotCurrentState(current_state)
}
\arguments{
\item{current_state}{Current-state table from \code{ssCompareCurrentState()}.}
}
\value{
ggplot object.
}
\description{
Plot current-state comparison using ggplot2
}
