% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{ssPlotHistorical}
\alias{ssPlotHistorical}
\title{Plot historical trend comparison using ggplot2}
\usage{
ssPlotHistorical(historical)
}
\arguments{
\item{historical}{Historical table from \code{ssCompareHistorical()}.}
}
\value{
ggplot object.
}
\description{
Plot historical trend comparison using ggplot2
}
