% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{ssPlotJabbaComparisons}
\alias{ssPlotJabbaComparisons}
\title{Build all SS-vs-JABBA comparison plots using ggplot2}
\usage{
ssPlotJabbaComparisons(comparisons)
}
\arguments{
\item{comparisons}{Named list from \code{ssCompareJabba()} or \code{ssJabbaWorkflow()$comparisons}.}
}
\value{
Named list of ggplot objects.
}
\description{
Build all SS-vs-JABBA comparison plots using ggplot2
}
