% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{ssPlotProduction}
\alias{ssPlotProduction}
\title{Plot production function comparison using ggplot2}
\usage{
ssPlotProduction(production)
}
\arguments{
\item{production}{Production table from \code{ssCompareProduction()}.}
}
\value{
ggplot object.
}
\description{
Plot production function comparison using ggplot2
}
