% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{ssPlotProjections}
\alias{ssPlotProjections}
\title{Plot projections (F=0 and F=FMSY) using ggplot2}
\usage{
ssPlotProjections(projections)
}
\arguments{
\item{projections}{Projection table from \code{ssCompareProjections()}.}
}
\value{
ggplot object.
}
\description{
Plot projections (F=0 and F=FMSY) using ggplot2
}
