% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-diags.R
\name{ssQqLine}
\alias{ssQqLine}
\title{Build QQ reference line coefficients}
\usage{
ssQqLine(x, y)
}
\arguments{
\item{x}{Numeric x values.}

\item{y}{Numeric y values.}
}
\value{
Named numeric vector with \code{a} and \code{b}.
}
\description{
Build QQ reference line coefficients
}
