% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssRun}
\alias{ssRun}
\title{Run SS executable in a model directory}
\usage{
ssRun(dir, ss3_exe = "ss3", nohess = TRUE)
}
\arguments{
\item{dir}{SS3 model directory.}

\item{ss3_exe}{Executable name/path.}

\item{nohess}{Logical; run with \code{-nohess}.}
}
\value{
Called for side effects; errors on non-zero exit.
}
\description{
Run SS executable in a model directory
}
