% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssRunSummary}
\alias{ssRunSummary}
\title{Summarize existing SS run outputs}
\usage{
ssRunSummary(ssDir, verbose = TRUE)
}
\arguments{
\item{ssDir}{SS3 model directory.}

\item{verbose}{Logical flag.}
}
\value{
List with run summary and run notes.
}
\description{
Summarize existing SS run outputs
}
