% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-read.R
\name{ssSlot}
\alias{ssSlot}
\title{Extract one slot from an \code{SS_output} object}
\usage{
ssSlot(rep, slot)
}
\arguments{
\item{rep}{\code{SS_output} list.}

\item{slot}{Slot name (case-insensitive), e.g. \code{"Kobe"}.}
}
\description{
Extract one slot from an \code{SS_output} object
}
