% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-retro-backtest.R
\name{tsDf}
\alias{tsDf}
\title{Timeseries data.frame from \code{r4ss::SS_output()}}
\usage{
tsDf(rep)
}
\arguments{
\item{rep}{A list returned by \code{SS_output()}.}
}
\value{
A single data.frame of the SS3 TIME_SERIES table.
}
\description{
Timeseries data.frame from \code{r4ss::SS_output()}
}
