% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-run.R
\name{readSs3Run}
\alias{readSs3Run}
\title{Read an SS3 report object from cache or Report.sso}
\usage{
readSs3Run(run_dir, save_rds = FALSE)
}
\arguments{
\item{run_dir}{SS3 run directory.}

\item{save_rds}{If \code{TRUE} and SS3 is run, save \code{ss.RData} after read.}
}
\description{
Loads \code{ss.RData} when present; otherwise calls \code{r4ss::SS_output()}.
}
