% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-jabba.R
\name{readTruth}
\alias{readTruth}
\alias{readTruth,character-method}
\title{Read SS3 truth object via ss3om}
\usage{
readTruth(x, ...)

\S4method{readTruth}{character}(x, ...)
}
\arguments{
\item{x}{SS3 directory path.}

\item{...}{Additional arguments passed to \code{ss3om::readFLSss3()}.}
}
\value{
FLR-compatible truth object.
}
\description{
Read SS3 truth object via ss3om
}
