% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudodataPipeline.R
\name{ssScenarios}
\alias{ssScenarios}
\title{Expand stock/scenario definitions into task list}
\usage{
ssScenarios(stocks)
}
\arguments{
\item{stocks}{List of stock specifications.}
}
\value{
List of task entries.
}
\description{
Expand stock/scenario definitions into task list
}
